Moved AggregateException to its own file.
